Use `DeferredCrateSelect` for CLI as well
